requirements.txt
# Mushroom LED Sculpture Dependencies
# Python 3.9+ required

# LED Control
rpi-ws281x==5.0.0

# Audio Processing
sounddevice==0.4.6
numpy==1.24.3
scipy==1.10.1

# Configuration & Utilities
pyyaml==6.0.1
psutil==5.9.5

# Development dependencies (optional)
# pytest==7.4.3
# black==23.12.0
# pylint==3.0.3